epwa2708it6/line02/requirements.txt
python-firebase==1.2
line-bot-sdk
Flask==1.0.2
